add gradle wrapper
Most tools require installation on your computer before you can use them.
If the installation is easy, you may think that’s fine.
But it can be an unnecessary burden on the users of the build.
Equally importantly, will the user install the right version of the tool
for the build? What if they’re building an old version of the software?

The Gradle Wrapper (henceforth referred to as the “Wrapper”) solves both
these problems and is the preferred way of starting a Gradle build.

more see https://docs.gradle.org/current/userguide/gradle_wrapper.html